NAMP add new libft

+++ nmap2me/libft/ft_printf/pf_u.c
@@ -3,18 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   pf_u.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: gbourgeo <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: gbourgeo <gbourgeo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/08/16 04:14:39 by gbourgeo          #+#    #+#             */
-/*   Updated: 2017/08/17 11:07:59 by gbourgeo         ###   ########.fr       */
+/*   Updated: 2020/02/20 21:37:28 by gbourgeo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "libft.h"
-#include <stdint.h>
 
-static ULL		get_modifier(t_dt *data)
+static unsigned long long		get_modifier(t_dt *data)
 {
 	if (*data->tail == 'U')
 		return (va_arg(data->ap, unsigned long));
@@ -33,7 +32,7 @@ static ULL		get_modifier(t_dt *data)
 	return (va_arg(data->ap, unsigned int));
 }
 
-void			pf_u(t_dt *data)
+void							pf_u(t_dt *data)
 {
 	t_av		av;
 	int			len;
